ajax

+++ app/Controllers/Http/Views/CorrectionController.js
@@ -25,41 +25,46 @@ class CorrectionController extends CorrectionSongcontroller{
 
     async createCorrection({ request, response}){
      
-        
+      if (request.ajax()) {
        const res =  await this.store({ request, response});
        return res;
+      }
       
     }
 
     async editCorrection({params, request, response}){
-      
+      if (request.ajax()) {
         const songctrl= new SongController();
          let res =  await songctrl.update ({ params, request, response });
          res = await this.destroy({request, response });
         return res;
+      }
       
      }
 
      async deleteCorrection({request,params, response}){
-    
+      if (request.ajax()) {
         const res =  await this.delete({ params, response });
         return res;
+      }
       
      }
 
 
     async getCorrection({request,response}){
-     
+      if (request.ajax()) {
         const res =  await this.index({response});
         return res;
+      }
       
      }
 
      async getCorrectionById({request,params, response}){
-    
+      if (request.ajax()) {
       const song = await Correction.find(params.id)
         
         return response.json({data:song});
+      }
       
      }
 }
